backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
openai==1.3.0
Pillow==10.1.0
pytest==7.4.3
